Add check before loading chartLoader to see if it is already loading/loaded from script tag.
Clean up imports.
