+++ demo/index.html
@@ -1,6 +1,8 @@
 <html>
 
 <head>
+  <script type="text/javascript" src="https://www.gstatic.com/charts/loader.js"></script>
+
 </head>
 
 <body>
